Fix some issues with ember-cp-validations
