--- tests/dummy/app/mixins/ember-cp-validations-helper.js
@@ -1,3 +1,0 @@
-import emberCpValidationsHelper from 'ember-rapid-forms/mixins/ember-cp-validations-helper';
-
-export default emberCpValidationsHelper;
