--- tests/dummy/app/routes/ember-cp-validations.js
@@ -1,13 +1,0 @@
-import Ember from 'ember';
-
-export default Ember.Route.extend({
-  model: function() {
-    return this.get('store').createRecord('person-with-cp-validations', {
-            name: "",
-            password: "",
-            comment: "",
-            active: false,
-            gender: "M"
-        });
-  }
-});
