+++ tests/dummy/app/routes/overview.js
@@ -1,13 +1,13 @@
 import Ember from 'ember';
 
 export default Ember.Route.extend({
-    model: function() {
-        return this.get('store').createRecord('simple_person', {
-            name: "",
-            password: "",
-            comment: "",
-            active: false,
-            gender: "M"
-        });
-    }
+  model: function() {
+    return this.get('store').createRecord('simple_person', {
+      name: "",
+      password: "",
+      comment: "",
+      active: false,
+      gender: "M"
+    });
+  }
 });
