--- tests/dummy/app/templates/ember-cp-validations.hbs
@@ -1,25 +1,0 @@
-<h1>Overview</h1>
-<p>
-<b>Ember Forms</b> is a small library for <i>Ember.js</i> that allows you to create astonishing forms
-    with support for <i> layouts, styles & validations</i> thanks to the combination of the beauty of <i>Bootstrap</i> and the power of <i>Ember</i>.
-</p>
-
-<hr/>
-
-<h4>And here is the result in a nutshell...</h4>
-
-<a href="http://jsbin.com/pexolude/201" target="_blank" class="btn btn-warning jsbin pull-right">jsBin</a>
-<div class="well line-example">
-    {{form-sample model=model}}
-</div>
-<div class="well line-example">
-    {{model.asjson}}
-</div>
-
-<h1>Want to see more?</h1>
-<p>
-    For getting started & installation steps, please visit {{#link-to 'getstarted'}}Getting Started{{/link-to}}
-</p>
-<p>
-    To see controls in action, visit {{#link-to 'controls'}}Controls{{/link-to}}
-</p>
